# Conflicts:
#	app/build.gradle
#	app/src/main/java/pt/it/porto/mydiabetes/adviceSystem/yapDroid/YapDroid.java
#	app/src/main/java/pt/it/porto/mydiabetes/data/Advice.java
#	app/src/main/java/pt/it/porto/mydiabetes/data/Task.java
#	app/src/main/java/pt/it/porto/mydiabetes/ui/activities/BaseActivity.java
#	app/src/main/java/pt/it/porto/mydiabetes/ui/activities/Home.java
#	app/src/main/java/pt/it/porto/mydiabetes/ui/fragments/home/homeMiddleFragment.java
#	build.gradle
#	gradle/wrapper/gradle-wrapper.properties

+++ app/src/main/java/pt/it/porto/mydiabetes/ui/charts/BodyOverlapHeaderGesture.java
@@ -29,7 +29,7 @@ public abstract class BodyOverlapHeaderGesture extends GestureDetector.SimpleOnG
 	private static final Property<View, Integer> VIEW_LAYOUT_HEIGHT = new Property<View, Integer>(Integer.class, "viewLayoutHeight") {
 
 		public void set(View object, Integer value) {
-			object.getLayoutParams().height = value.intValue();
+			object.getLayoutParams().height = value;
 			object.requestLayout();
 		}
